change to pep8 style, add item json endpoint, add verification for edit/delete item

+++ project.py
@@ -1,5 +1,5 @@
 from flask import Flask, render_template, request, redirect, jsonify, url_for, flash
-from sqlalchemy import create_engine, asc
+from sqlalchemy import create_engine, asc, and_
 from sqlalchemy.orm import sessionmaker
 from database_setup import Base, Category, Item, User
 from flask import session as login_session
@@ -12,9 +12,9 @@ import json
 from flask import make_response
 import requests
 
-app = Flask(__name__,static_folder='/vagrant/static')
+app = Flask(__name__, static_folder='/vagrant/static')
 # Connect to Database and create database session
-engine = create_engine('postgresql://postgres:postgres@localhost/news')
+engine = create_engine('postgresql://catalog:password@localhost/catalog')
 Base.metadata.bind = engine
 
 DBSession = sessionmaker(bind=engine)
@@ -25,17 +25,27 @@ CLIENT_ID = json.loads(
     open('client_secrets.json', 'r').read())['web']['client_id']
 APPLICATION_NAME = "Restaurant Menu Application"
 
-#json part
+# json part
+
+
 @app.route('/catalog.json')
 def restaurantsJSON():
     catagories = session.query(Category).all()
     tempdata = [r.serialize for r in catagories]
     for cat in tempdata:
-        cat["Item"] =[x.serialize for x in session.query(Item).filter_by(catagory_id = cat["id"]).all()]
+        cat["Item"] = [x.serialize for x in session.query(
+            Item).filter_by(catagory_id=cat["id"]).all()]
     return jsonify(Category=tempdata)
 
 
+@app.route('/catalog/<int:catid>/<int:itemid>/json')
+def itemjson(catid, itemid):
+    item = session.query(Item).filter_by(id=itemid).one()
+    return jsonify(Item=item.serialize)
+
 # Create anti-forgery state token
+
+
 @app.route('/login')
 def showLogin():
     state = ''.join(random.choice(string.ascii_uppercase + string.digits)
@@ -43,6 +53,7 @@ def showLogin():
     login_session['state'] = state
     # return "The current session state is %s" % login_session['state']
     return render_template('login.html', STATE=state)
+
 
 @app.route('/gconnect', methods=['POST'])
 def gconnect():
@@ -58,7 +69,8 @@ def gconnect():
         # Upgrade the authorization code into a credentials object
         oauth_flow = flow_from_clientsecrets('client_secrets.json', scope='')
         oauth_flow.redirect_uri = 'postmessage'
-        credentials = oauth_flow.step2_exchange(code)#use code from google get user's access token
+        # use code from google get user's access token
+        credentials = oauth_flow.step2_exchange(code)
     except FlowExchangeError:
         response = make_response(
             json.dumps('Failed to upgrade the authorization code.'), 401)
@@ -115,14 +127,13 @@ def gconnect():
     login_session['username'] = data['name']
     login_session['picture'] = data['picture']
     login_session['email'] = data['email']
-    #see if a user exists
+    # see if a user exists
     user_id = getUserID(login_session['email'])
     print user_id
     if not user_id:
         user_id = createUser(login_session)
         print user_id
     login_session['user_id'] = user_id
-
 
     output = ''
     output += '<h1>Welcome, '
@@ -135,8 +146,10 @@ def gconnect():
     print "done!"
     return output
 
-#create a process to disconnect the page
-@app.route('/gdisconnect')#use google api to revoke token
+# create a process to disconnect the page
+
+
+@app.route('/gdisconnect')  # use google api to revoke token
 def gdisconnect():
     access_token = login_session.get('access_token')
     if access_token is None:
@@ -144,8 +157,8 @@ def gdisconnect():
         response = make_response(json.dumps('current user not connected'), 401)
         response.headers['Content-Type'] = 'application/json'
         return response
-    #print 'revoken token is {}'.format(access_token)
-    #print 'user name is {}'.format(login_session['username'])
+    # print 'revoken token is {}'.format(access_token)
+    # print 'user name is {}'.format(login_session['username'])
     url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' % login_session['access_token']
     h = httplib2.Http()
     result = h.request(url, 'GET')[0]
@@ -160,29 +173,41 @@ def gdisconnect():
         response.headers['Content-Type'] = 'application/json'
         return response
     else:
-        response = make_response(json.dumps('Failed to revoke token for given user.', 400))
+        response = make_response(
+            json.dumps(
+                'Failed to revoke token for given user.',
+                400))
         response.headers['Content-Type'] = 'application/json'
         return response
 
-#restrict page by user
+# restrict page by user
+
+
 def createUser(login_session):
-    newUser = User(name = login_session['username'], email = login_session['email'], picture = login_session['picture'])
+    newUser = User(
+        name=login_session['username'],
+        email=login_session['email'],
+        picture=login_session['picture'])
     session.add(newUser)
     session.commit()
-    user = session.query(User).filter_by(email = login_session['email']).one()
+    user = session.query(User).filter_by(email=login_session['email']).one()
     return user.id
 
+
 def getUserInfo(user_id):
-    user = session.query(User).filter_by(id = user_id).one()
+    user = session.query(User).filter_by(id=user_id).one()
     return user
+
 
 def getUserID(email):
     try:
-        user = session.query(User).filter_by(email = email).one()
+        user = session.query(User).filter_by(email=email).one()
         return user.id
-    except:
+    except BaseException:
         return None
 # Disconnect based on provider
+
+
 @app.route('/disconnect')
 def disconnect():
     print login_session
@@ -194,19 +219,23 @@ def disconnect():
     #del login_session['picture']
     #del login_session['user_id']
     #del login_session['provider']
-    
-    return redirect(url_for('showcatagories'))
-    
 
-#show all categoryies
+    return redirect(url_for('showcatagories'))
+
+
+# show all categoryies
 @app.route('/')
 @app.route('/catagory/')
 def showcatagories():
     catagories = session.query(Category).all()
     items = session.query(Item).order_by(Item.created_date.desc()).all()
-    itemswithcat = [session.query(Category).filter_by(id=item.catagory_id).one() for item in items]
-    print(session.query(Category).filter_by(id=items[0].catagory_id).one().name)
-    return render_template('catagories.html', catagories = catagories, items = zip(items, itemswithcat))
+    itemswithcat = [session.query(Category).filter_by(
+        id=item.catagory_id).one() for item in items]
+    print(session.query(Category).filter_by(
+        id=items[0].catagory_id).one().name)
+    return render_template(
+        'catagories.html', catagories=catagories, items=zip(items, itemswithcat))
+
 
 @app.route('/catagory/<string:catagory>/items/')
 @app.route('/catagory/<string:catagory>/')
@@ -216,29 +245,36 @@ def showItem(catagory):
     items = session.query(Item).filter_by(catagory_id=catagory.id).all()
     creator = catagory.user_id
     total = len(items)
-    
+
     if 'username' not in login_session:
-        return render_template('publicitem.html', catagories = catagories, items = items, catagory = catagory, total = total)
+        return render_template('publicitem.html', catagories=catagories,
+                               items=items, catagory=catagory, total=total)
     else:
-        return render_template('item.html', catagories = catagories, items = items, catagory = catagory, total = total)
+        return render_template(
+            'item.html', catagories=catagories, items=items, catagory=catagory, total=total)
 
 
 @app.route('/catagory/<string:catagory_name>/<string:item_name>/')
 def showDescription(catagory_name, item_name):
-    #print login_session['user_id']
+    # print login_session['user_id']
     item = session.query(Item).filter_by(name=item_name).one()
     catagory = session.query(Category).filter_by(name=catagory_name).one()
     creator_id = item.user_id
     if login_session['user_id'] == creator_id:
 
-        return render_template('description.html', item = item, catagory = catagory)
+        return render_template(
+            'description.html', item=item, catagory=catagory)
     else:
-        return render_template('publicdescription.html', item = item, catagory = catagory)
+        return render_template('publicdescription.html',
+                               item=item, catagory=catagory)
+
 
 @app.route('/catagory/<string:item>/edit', methods=['GET', 'POST'])
 def editItem(item):
+    if 'username' not in login_session:
+        return redirect('/login')
     item = session.query(Item).filter_by(name=item).one()
-    catagory = session.query(Category).filter_by(id = item.catagory_id).one()
+    catagory = session.query(Category).filter_by(id=item.catagory_id).one()
     total_cat = session.query(Category).all()
     if request.method == 'POST':
         if request.form['name']:
@@ -246,32 +282,45 @@ def editItem(item):
         if request.form['description']:
             item.description = request.form['description']
         if request.form['catagory']:
-            cat_id = session.query(Category).filter_by(name = request.form['catagory']).one().id
-            #print(cat_id)
+            cat_id = session.query(Category).filter_by(
+                name=request.form['catagory']).one().id
+            # print(cat_id)
             item.catagory_id = int(cat_id)
         session.add(item)
         session.commit()
         flash('Item Successfully Edited')
         return redirect(url_for('showItem', catagory=catagory.name))
     else:
-        return render_template('edit.html', item = item, catagory = catagory, total_cat = total_cat)
+        return render_template('edit.html', item=item,
+                               catagory=catagory, total_cat=total_cat)
+
 
 @app.route('/catagory/<string:item>/delete', methods=['GET', 'POST'])
 def deleteItem(item):
+    if 'username' not in login_session:
+        return redirect('/login')
     item = session.query(Item).filter_by(name=item).one()
-    catagory = session.query(Category).filter_by(id = item.catagory_id).one()
+    catagory = session.query(Category).filter_by(id=item.catagory_id).one()
     if request.method == 'POST':
         session.delete(item)
         session.commit()
         return redirect(url_for('showItem', catagory=catagory.name))
     else:
         return render_template('delete.html')
+
+
 @app.route('/catagory/<string:catagory_name>/additem', methods=['GET', 'POST'])
 def addItem(catagory_name):
-    catagory = session.query(Category).filter_by(name = catagory_name).one()
+    if 'username' not in login_session:
+        return redirect('/login')
+    catagory = session.query(Category).filter_by(name=catagory_name).one()
     if request.method == 'POST':
         if request.form['name'] and request.form['description']:
-            newitem = Item(name = request.form['name'], description = request.form['description'], catagory_id = catagory.id, user_id = login_session['user_id'])
+            newitem = Item(
+                name=request.form['name'],
+                description=request.form['description'],
+                catagory_id=catagory.id,
+                user_id=login_session['user_id'])
             session.add(newitem)
             session.commit()
             flash('Item Successfully added')
